requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
Django==5.0.4
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.14.0
docutils==0.21.2
drf-spectacular==0.27.2
flake8==7.0.0
flake8-docstrings==1.7.0
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
packaging==24.0
pillow==10.3.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sentry-sdk==2.3.1
snowballstemmer==2.2.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
